Initial commit.
Initial commit of the utilities parent project which contains the file-renamer Maven Plugin project.

file-renamer created to allow the renaming of a file in the 'package' phase. This need arose when the nar-maven and maven-bundle plugins would not work together and allow the creation of a bundle containing native libraries named '.jar' but instead names them ending '.nar'. The default can be overridden for other file types.